Define view_live_pages once in base, reuse everywhere else. I think this is the least flakey version
